+++ activerecord/test/cases/calculations_test.rb
@@ -493,6 +493,12 @@ class CalculationsTest < ActiveRecord::TestCase
     assert_equal [1,2,3,4], Topic.order(:id).pluck("topics.id")
   end
 
+  def test_pluck_replaces_select_clause
+    taks_relation = Topic.select([:approved, :id]).order(:id)
+    assert_equal [1,2,3,4], taks_relation.pluck(:id)
+    assert_equal [false, true, true, true], taks_relation.pluck(:approved)
+  end
+
   def test_pluck_auto_table_name_prefix
     c = Company.create!(:name => "test", :contracts => [Contract.new])
     assert_equal [c.id], Company.joins(:contracts).pluck(:id)